refactor: remove tmpfile usage.
Removes dependency of the `tmpfile` builtin. The reason for this is the
tmpfile was being created when downloading files from S3, and then being
deleted. However, because tmpfiles have a random component in the name,
two things were needed to be kept track of in GeoFile; the "path" name
(with the random component) and the actual file name. Since there was no
actual benefit of using a tempfile since the deletion has to be done
manually anyway to provide support on Windows as well, this process was
removed and the file is saved directly to the path specified by the
user.